+++ java-client-sdk/lib/src/main/java/industries/dot/brease/SDK.java
@@ -35,8 +35,8 @@ public class SDK {
 	
 	private String _serverUrl;
 	private String _language = "java";
-	private String _sdkVersion = "0.2.2";
-	private String _genVersion = "2.26.4";
+	private String _sdkVersion = "0.3.0";
+	private String _genVersion = "2.27.0";
 	/**
 	 * The Builder class allows the configuration of a new instance of the SDK.
 	 */