+++ python-client-sdk/src/sdk/sdk.py
@@ -21,8 +21,8 @@ class SDK:
     _security_client: requests_http.Session
     _server_url: str = SERVERS[0]
     _language: str = "python"
-    _sdk_version: str = "1.1.1"
-    _gen_version: str = "2.26.4"
+    _sdk_version: str = "1.2.0"
+    _gen_version: str = "2.27.0"
 
     def __init__(self,
                  server_url: str = None,