+++ RELEASES.md
@@ -132,3 +132,13 @@ Based on:
 - [PyPI v1.1.1] https://pypi.org/project/brease-sdk/1.1.1 - python-client-sdk
 - [NPM v0.1.3] https://www.npmjs.com/package/brease-sdk/v/0.1.3 - typescript-client-sdk
 - [Go v0.3.2] https://github.com/speakeasy-sdks/brease/releases/tag/go-client-sdk/v0.3.2 - go-client-sdk
+
+## 2023-05-13 00:39:38
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 https://prod.speakeasyapi.dev/v1/apis/brease/version/0.1/schema/download
+- Speakeasy CLI 1.31.1 (2.27.0) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.4.0] https://github.com/speakeasy-sdks/brease/releases/tag/go-client-sdk/v0.4.0 - go-client-sdk
+- [PyPI v1.2.0] https://pypi.org/project/brease-sdk/1.2.0 - python-client-sdk
+- [NPM v0.2.0] https://www.npmjs.com/package/brease-sdk/v/0.2.0 - typescript-client-sdk